Launcher3: Change resource type for blur radius in DepthController
This allows us to define the blur radius in device-independent pixels
instead of physical pixels, which is necessary to fix discrepancies in
apparent blur strength across devices with different pixel densities.

+++ quickstep/src/com/android/quickstep/util/BaseDepthController.java
@@ -63,7 +63,7 @@ public class BaseDepthController {
     /**
      * Blur radius when completely zoomed out, in pixels.
      */
-    protected final int mMaxBlurRadius;
+    protected final float mMaxBlurRadius;
     protected final WallpaperManager mWallpaperManager;
     protected boolean mCrossWindowBlursEnabled;
